backend/requirements.txt
appdata==2.1.2
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
bs4==0.0.1
certifi @ file:///croot/certifi_1671487769961/work/certifi
click==8.0.3
colored==1.4.3
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
lxml==4.9.2
pytz==2022.7.1
soupsieve==2.4
sqlparse==0.4.3
